- Users subsystem fixes
