+++ src/services/auth.service.js
@@ -1,6 +1,7 @@
 import StitchService from "./stitch.service";
 import EventEmitter from "events";
 import storage from '../helpers/localStorageData';
+import {checkUserRole} from '../helpers/get-data';
 
 const stitchService = StitchService.getInstance();
 
@@ -14,17 +15,25 @@ export default class AuthService extends EventEmitter {
     return AuthService.serviceInstance;
   }
 
+  get userRole(){
+    return this.user ? checkUserRole(this.user) : "unauthenticated"
+  }
+
   get isStitchAuthenticated() {
     return storage.getItem("__stitch.client." + stitchService.appId + ".auth_info");
   }
 
   get isAuthenticated() {
-    return this.user != null;
+    return this._user != null;
+  }
+
+  get user(){
+    return this._user ? this._user.customData : this._user;
   }
 
   authenticate(login, pass) {
     return stitchService.authenticateStitch(login, pass).then(user => {
-      this.user = user;
+      this._user = user;
       delete (user.auth);
       storage.setAuthInfo(user);
       this.emit("authorized", user);
@@ -33,6 +42,6 @@ export default class AuthService extends EventEmitter {
 
   getUserFromLocalStorage(user) {
     stitchService.reinitializeClient();
-    return this.user = storage.getAuthInfo();
+    return this._user = storage.getAuthInfo();
   }
 }
