+++ src/services/user.service.js
@@ -32,6 +32,7 @@ export default class UserService {
   }
 
   createUser(password, data) {
+    debugger;
     return stitchService.client.auth
       .getProviderClient(UserPasswordAuthProviderClient.factory)
       .registerWithEmail(data.email, password)
